% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSampleSheet.R
\name{importSampleSheet}
\alias{importSampleSheet}
\title{importSampleSheet}
\usage{
importSampleSheet(sampleSheet = NULL, well_id = NULL)
}
\arguments{
\item{sampleSheet}{Path to sample sheet file containg information about
samples.}

\item{well_id}{Character vector with well IDs corresponding to uploaded
amplitude files.}
}
\value{
A matrix with columns for sample ID, target assay and control assay.
}
\description{
Function that takes a path to a .csv-file containing information
    about the samples that correspond to the uploaded amplitude files. This
    file must contain the following columns: Well, Sample, TargetType
    and Target. A character vector with well IDs must also be provided,
    which is used to match rows in sample sheet to amplitude files
}
\examples{
## Path to example sample sheet included in PoDCall
path <- system.file("extdata", "Sample_names.csv", package="PoDCall")

## Select wells to get information for
well_id <- c("A04", "B04", "D04")

## Get information for selected wells
sampleSheet <- importSampleSheet(sampleSheet=path, well_id=well_id)

}
