% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPoDCall.R
\name{podcall_ddpcr}
\alias{podcall_ddpcr}
\title{Positive Droplet Calling for ddPCR}
\usage{
podcall_ddpcr(
  data_directory,
  ss_filename = NULL,
  B = 200,
  Q = 7.5,
  refwell = 1,
  ch2 = TRUE,
  results_to_file = FALSE,
  plots = FALSE
)
}
\arguments{
\item{data_directory}{Directory containing Quantasoft amplitude files}

\item{ss_filename}{File (optional) containing sample information from ddPCR experiment}

\item{B}{The number of permutations used for the LRT (default=200)}

\item{Q}{A parameter for calling outliers (default=7.5)}

\item{refwell}{reference well to calculate the shift in baseline (default=1)}

\item{ch2}{Logical argument to denote channel 2 amplitudes (default=TRUE)}

\item{results_to_file}{Should results be written to file (.csv)? (default=FALSE)}

\item{plots}{Should plots be created and written to file? (default=FALSE)}
}
\value{
The function returns a table with thresholds, droplet counts, concentration and normalized concentration.
    The table is also written to a .csv-file and plots for both channels per well is written to files (unless suppressed by user).
}
\description{
Function that takes path to amplitude files and parameters for setting threshold as input.
    Calls functions that read in data from files, sets threshold for each channel per well, calculates
    concentrations and optionally makes scatterplot and histogram for each channel per well.
    Results are returned as a table, optionally written to file. Plots will only be made if user
    chooses to save results to file.
}
\examples{
\dontrun{
podcall_ddpcr(data_directory="datapath")
}
}
