% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podcallShiny.R
\name{podcallShiny}
\alias{podcallShiny}
\title{PoDCall shiny launcher}
\usage{
podcallShiny()
}
\value{
Does not return anything, but launches PoDCall shiny app
}
\description{
This function launches the PoDCall shiny app in a web browser
}
\examples{
\dontrun{
podcallShiny()
}
}
