% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPoDCall.R
\name{podcall_ddpcr}
\alias{podcall_ddpcr}
\title{Positive Droplet Calling for ddPCR}
\usage{
podcall_ddpcr(
  dataDirectory,
  sampleSheetFile = NULL,
  B = 400,
  Q = 9,
  refwell = 1,
  ch2 = TRUE,
  resultsToFile = FALSE,
  plots = FALSE
)
}
\arguments{
\item{dataDirectory}{Path to directory containing Quantasoft amplitude files
from one 96 well plate. Since well coordinates are used as identifiers,
files in this directory should all be from the same 96 well plate.
Furthermore, there can be no other files than the amplitude files from
a well plate in the directory.}

\item{sampleSheetFile}{File (optional) containing sample information from
ddPCR experiment. This file must be a comma separated file containing the
following columns: Well, Sample, TargetType and Target.}

\item{B}{The number of permutations used for the Likelihood Ratio Test
(default=400)}

\item{Q}{A parameter for calling outliers (default=9)}

\item{refwell}{reference well to calculate the shift in baseline (default=1)}

\item{ch2}{Logical argument to denote channel 2 amplitudes (default=TRUE)}

\item{resultsToFile}{Should results be written to file(.csv)? (default=FALSE)}

\item{plots}{Should plots be created and written to file? (default=FALSE)}
}
\value{
The function returns a table (data frame) with thresholds,
    droplet counts, concentration and normalized concentration.
    The table is optionally written to a .csv-file and plots for both
    channels per well can be written to files.
}
\description{
Wrapper function that provide a complete workflow for the
    functionality of PoDCall. It takes path to amplitude files and sample
    sheet (optional), and parameters for setting threshold as input.
    Calls functions that read in data from files, sets threshold for each
    channel per well, calculates concentrations and optionally makes
    scatterplot and histogram for each channel per well.
    Results are returned as a table, optionally written to file.
    Plots will only be made if user chooses to save results to file, and will
    be written to file in a results directory.
}
\examples{
\dontrun{
# Paths to data and sample sheet
dataPath <- system.file("extdata", "Amplitudes/", package="PoDCall")
ssPath <- system.file("extdata", "Sample_names.csv", package="PoDCall")

# Run PodCall
podcallResults <- podcall_ddpcr(dataDirectory=dataPath,
                                 sampleSheetFile=ssPath)
}
}
