% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/podcallThresholds.R
\name{podcallThresholds}
\alias{podcallThresholds}
\title{podcallThresholds}
\usage{
podcallThresholds(plateData,
                        nchannels=c(1,2)[2],
                        B=200,
                        Q=9,
                        refWell=1,
                        updateProgress=NULL)
}
\arguments{
\item{plateData}{List of data frames with amplitude data from a 96 well plate}

\item{nchannels}{Number of channels used in the experiment (default=2)}

\item{B}{Number of permutations for the Likelihood Ratio Test (LRT)
(default=200)}

\item{Q}{Parameter for outlier calling (default=9)}

\item{refWell}{reference well to calculate the shift in baseline (default=1)}

\item{updateProgress}{function to update progress bar in shiny app
(default=NULL)}
}
\value{
A table with results and metrics, one row per well.
}
\description{
Function sets threshold per channel per well and calculates
    concentrations. Results are returned as a data frame.
}
\examples{
## Path to example data
dataPath <- system.file("extdata", "Amplitudes/", package="PoDCall")

## Read in example data
dataList <- importAmplitudeData(dataDirectory=dataPath)

## Set thresholds
thresholds <- podcallThresholds(plateData=dataList,
                                B=200)

}
